requirements.txt
Django>=5.2
djangorestframework
mongoengine
shortuuid
requests
user_agents
datetime


